+++ scripts/lib/coaching.sh
@@ -32,6 +32,10 @@ coaching_build_behavior_digest() {
     coach_build_behavior_digest "$@"
 }
 
+coaching_collect_local_context_bundle() {
+    coach_collect_local_context_bundle "$@"
+}
+
 coaching_build_prebrief_questions() {
     coach_build_prebrief_questions "$@"
 }